Add methods to detect touch locations
These allow the game to define a region of interest for a touch event
and then detect the "most important" finger in the region. Also allows
for a "secondary" finger to be used which will ignore a finger which has
already been used for some other input (or to ignore regions).