Initial implementation of a TOUCH mode
Plays like TWIN_STICK with two virtual joysticks in the lower right and
left corners of the screen. The left controls movement and the right
controls the direction of fire.

Things to fix: the size and location of the input areas and the input
locations need to be drawn onto the screen. TODO items mark the places
these should happen.